[BUG-111075][CALCITE-2469] RexSimplify should optimize '(NOT x) IS NULL' to 'x IS NULL' (pengzhiwei)
Previously it optimized '(NOT x) IS NULL' to 'x IS NOT NULL', which is
wrong.

Generalize the above, to simplify 'f(x) IS NULL' to 'x IS NULL' for any
operator 'f' that is known to be strong. (Julian Hyde)

Close apache/calcite#796

(cherry picked from commit 90f49be639b9a569e3dc4d8369c08e2db64ea301)

Change-Id: I8b5a1b2f0f9890aca561a5cf4a5c3ec0013bc7bb